Update main --all.pyw

+++ main --all.pyw
@@ -1,3 +1,12 @@
+#print(ad) || total
+#print(ad[0]) || Name
+#print(ad[1]) || Passhashed sha3_512
+#print(ad[2]) || Wins
+#print(ad[3]) || Losses
+#print(ad[4]) || Draws
+#print(ad[5]) || Account type || 0/Standard | 1/SuperUser | 2/Admin | 3/Debug
+#print(ad[6]) || Profile Picture Location
+
 #Imports
 import tkinter as tk
 from PIL import ImageTk, Image
@@ -33,16 +42,24 @@ class startMenu(tk.Frame):
 
     def playSingle(self):
         startwindow.destroy()
-        game.singlePlayer()
+        try:
+            profilewindow.destroy()
+        except:
+            pass
+        #game.singlePlayer()
         
     def playMulti(self):
         startwindow.destroy()
-        game.multiPlayer()
+        try:
+            profilewindow.destroy()
+        except:
+            pass
+        #game.multiPlayer()
         
     def profile(self):
         if currentUser != '':
             global profilewindow
-            profilewindow = tk.Tk()
+            profilewindow = tk.Toplevel()
             profileMenu(profilewindow).pack(fill="both", expand=True)
             profilewindow.mainloop()
         else:
@@ -110,14 +127,16 @@ class loginMenu(tk.Frame):
         try:
             os.mkdir(usrDirectory)
             passvarHashed = hashlib.sha3_512(passvar.encode('utf-8')).hexdigest()
+            global currentUser
             currentUser = str(usrvar)
             del usrvar
             del passvar
             initWins = 0
             initLoses = 0
+            initDraws = 0
             initAccountType = 0
             initProfilePicture = "./lib/Blank-Profile-Picture.pgm"
-            usrVariables = [currentUser, passvarHashed, initWins, initLoses, initAccountType, initProfilePicture]
+            usrVariables = [currentUser, passvarHashed, initWins, initLoses, initDraws, initAccountType, initProfilePicture]
             DataLocation = ("./usr/%s/AccountDetails.dat" % currentUser)
             with open(DataLocation, "wb+") as data:
                 pickle.dump(usrVariables, data, 0)
@@ -128,8 +147,14 @@ class loginMenu(tk.Frame):
             self.subtext = tk.Label(self, text="You may now close this window", bg="white", font="TkFixedFont 12")
             global Wins
             global Loses
+            global Draws
+            global profilePicture
+            global accountType
             Wins = initWins
             Loses = initLoses
+            Draws = initDraws
+            accountType = initAccountType
+            profilePicture = initProfilePicture
             self.closeWindowButton = tk.Button(self, text="Close", bg="white", font="TkFixedFont 12", command=loginwindow.destroy)
             self.maintext.grid(column=0, row=0)
             self.subtext.grid(column=0, row=1)
@@ -157,25 +182,28 @@ class loginMenu(tk.Frame):
         global usrvar
         usrvar = self.userInputForm.get()
         DataLocation = ("./usr/%s" % usrvar)
-        try:
-            os.mkdir(DataLocation)
-            os.rmdir(DataLocation)
-            self.usrNotExists = tk.Label(self, text="User does not exists", bg="red")
-            self.usrNotExists.grid(column=1, row=5, sticky="s")
-        except FileExistsError:
-            DataLocation = ("./usr/%s/AccountDetails.dat" % usrvar)
-            for widget in self.winfo_children():
-                widget.destroy()                  
-            self.config(bg="white")
-            self.title = tk.Label(self, text="Welcome %s" % usrvar, font="TkFixedFont 14", bg="white", anchor="n")
-            self.subtitle = tk.Label(self, text="Enter your password:", font="TkFixedFont 12", bg="white", anchor="n")
-            self.passwordInputForm = tk.Entry(self, show="*")
-            self.passwordInputForm.config(highlightbackground="white", highlightthickness=0)
-            self.credentialsSubmitButton = tk.Button(self, text="Login", bg="white", font="TkFixedFont 12", command=self.loginpasswordsubmit)
-            self.title.grid(column=1, row=1, sticky="n")
-            self.subtitle.grid(column=1, row=2, sticky="n")
-            self.passwordInputForm.grid(column=1, row=3, padx=20, sticky="e")
-            self.credentialsSubmitButton.grid(column=1, row=4, sticky="s")
+        if len(usrvar) > 0:
+            try:
+                os.mkdir(DataLocation)
+                os.rmdir(DataLocation)
+                self.usrNotExists = tk.Label(self, text="User does not exists", bg="red")
+                self.usrNotExists.grid(column=1, row=6, sticky="s")
+            except FileExistsError:
+                for widget in self.winfo_children():
+                    widget.destroy()                  
+                self.config(bg="white")
+                self.title = tk.Label(self, text="Welcome %s" % usrvar, font="TkFixedFont 14", bg="white", anchor="n")
+                self.subtitle = tk.Label(self, text="Enter your password:", font="TkFixedFont 12", bg="white", anchor="n")
+                self.passwordInputForm = tk.Entry(self, show="*")
+                self.passwordInputForm.config(highlightbackground="white", highlightthickness=0)
+                self.credentialsSubmitButton = tk.Button(self, text="Login", bg="white", font="TkFixedFont 12", command=self.loginpasswordsubmit)
+                self.title.grid(column=1, row=1, sticky="n")
+                self.subtitle.grid(column=1, row=2, sticky="n")
+                self.passwordInputForm.grid(column=1, row=3, padx=20, sticky="e")
+                self.credentialsSubmitButton.grid(column=1, row=4, sticky="s")
+        else:
+            self.usrBlank = tk.Label(self, text="Username cannot be Blank", bg="red")
+            self.usrBlank.grid(column=1, row=5, sticky="s")
             
     def loginpasswordsubmit(self):
         passvar = self.passwordInputForm.get()
@@ -190,12 +218,14 @@ class loginMenu(tk.Frame):
                 widget.destroy()
             global Wins
             global Loses
+            global Draws
             global accountType
-            global ProfilePicture
+            global profilePicture
             Wins = logonDetails[2]
             Loses = logonDetails[3]
-            accountType = logonDetails[4]
-            profilePicture = logonDetails[5]
+            Draws = logonDetails[4]
+            accountType = logonDetails[5]
+            profilePicture = logonDetails[6]
             self.config(bg="white")
             self.title = tk.Label(self, text="Welcome %s" % currentUser, font="TkFixedFont 14", bg="white", anchor="n")
             self.subtitle = tk.Label(self, text="Your Infomation has been accepted", font="TkFixedFont 12", bg="white", anchor="n")
@@ -209,25 +239,120 @@ class loginMenu(tk.Frame):
             self.passNotValid = tk.Label(self, text="Password is not valid", bg="red")
             self.passNotValid.grid(column=1, row=5, sticky="s")
         
-                #print(ad) || total
-                #print(ad[0]) || Name
-                #print(ad[1]) || Passhashed sha3_512
-                #print(ad[2]) || Wins
-                #print(ad[3]) || Losses
-                #print(ad[4]) || Account type || 0/Standard | 1/SuperUser | 2/Admin | 3/Debug
-                #print(ad[5]) || Profile Picture Location
-
 #Profile class
 class profileMenu(tk.Frame):
     def __init__(self, parent):
         tk.Frame.__init__(self, parent)
         self.config(bg="white")
-        profilewindow.title("%s's Account" % currentUser)
-        #self.logo = ImageTk.PhotoImage(Image.open(profilePicture))
-        self.title = tk.Label(self, text="Welcome %s" % currentUser, anchor="c", font="TkFixedFont 18 bold", bg="white")
-        self.subtitle = tk.Label(self, text="Made with Python 3.7 by Sasith De Abrew", bg="white", anchor="s")
-        self.title.pack(side="top", fill="x", expand=1)
-        self.subtitle.pack(side="bottom", fill="x", anchor="s")
+        profilewindow.title("%s's Profile" % currentUser)
+        self.title = tk.Label(self, text="Welcome %s" % currentUser, font="TkFixedFont 18 bold", bg="white")
+        self.logo = ImageTk.PhotoImage(Image.open(profilePicture))
+        self.logoLabel = tk.Label(self, image=self.logo, bg="white")
+        self.statsDisplay = tk.Label(self, text="Wins: %s\nLoses: %s\nDraws: %s" % (Wins, Loses, Draws), font="TkFixedFont 15 bold", bg="white")
+        self.editProfileButton = tk.Button(self, text="Edit Profile", bg="white", font="TkFixedFont 12", command=self.editprofile)    
+        self.exitProfileButton = tk.Button(self, text="Exit", bg="white", font="TkFixedFont 12", command=profilewindow.destroy)
+        self.title.grid(column=0, row=0, columnspan=2)
+        self.logoLabel.grid(column=0, row=1, sticky="w")
+        self.statsDisplay.grid(column=1, row=1, sticky="w")
+        self.editProfileButton.grid(column=0, row=2, sticky="news")
+        self.exitProfileButton.grid(column=1, row=2, sticky="news")
+        
+    def editprofile(self):
+        for widget in self.winfo_children():
+            widget.destroy()
+        self.title = tk.Label(self, text="Change account details", bg="white", font="TkFixedFont 15 bold")
+        self.changePictureButton = tk.Button(self, text="Change profile picture", bg="white", command=self.changeprofilepicture)
+        self.changeUsernameButton = tk.Button(self, text="Change account username", bg="white", command=self.changeaccountusername)
+        self.changePasswordButton = tk.Button(self, text="Change account password", bg="white", command=self.changeaccountpassword)
+        self.title.grid(column=0, row=0)
+        self.changePictureButton.grid(column=0, row=1, sticky="news")
+        self.changeUsernameButton.grid(column=0, row=2, sticky="news")
+        self.changePasswordButton.grid(column=0, row=3, sticky="news")
+
+    def changeprofilepicture(self):
+        print("placeholder")
+
+    def changeaccountusername(self):
+        for widget in self.winfo_children():
+            widget.destroy()
+        self.title = tk.Label(self, text="Change your username", bg="white", font="TkFixedFont 16 bold")
+        self.passwordSubtitle = tk.Label(self, text="Current password:", bg="white")
+        self.passwordConfirmSubtitle = tk.Label(self, text="Confirm password:", bg="white")
+        self.passwordInputForm = tk.Entry(self, show="*")
+        self.passwordConfirmForm = tk.Entry(self, show="*")
+        self.passwordSubmitButton = tk.Button(self, text="Submit", bg="white", command=self.changeuser)
+        self.title.grid(column=0, row=0, sticky="news", columnspan=2)
+        self.passwordSubtitle.grid(column=0, row=1)
+        self.passwordInputForm.grid(column=1, row=1)
+        self.passwordConfirmSubtitle.grid(column=0, row=2)
+        self.passwordConfirmForm.grid(column=1, row=2)
+        self.passwordSubmitButton.grid(column=0, row=3, sticky="news", columnspan=2)
+
+    def changeuser(self):
+        if self.passwordInputForm.get() == self.passwordConfirmForm.get():
+            DataLocation = ("./usr/%s/AccountDetails.dat" % currentUser)
+            with open(DataLocation, "rb+") as UsrData:
+                Details = pickle.load(UsrData)
+                UsrData.close()
+            if Details[1] == hashlib.sha3_512(self.passwordInputForm.get().encode("utf-8")).hexdigest() and Details[1] == hashlib.sha3_512(self.passwordConfirmForm.get().encode("utf-8")).hexdigest():
+                for widget in self.winfo_children():
+                    widget.destroy()
+                self.title = tk.Label(self, text="Change your username", bg="white", font="TkFixedFont 16 bold")
+                self.currentUserSubtitle = tk.Label(self, text="Current username:", bg="white")
+                self.currentUserForm = tk.Entry(self)
+                self.newUserSubtitle = tk.Label(self, text="New username:", bg="white")
+                self.newUserForm = tk.Entry(self)
+                self.usernameChangeConfirm = tk.Button(self, text="Confirm Changes", bg="white", command=self.confirmusernamechange)
+                self.title.grid(column=0, row=0, columnspan=2)
+                self.currentUserSubtitle.grid(column=0, row=1)
+                self.currentUserForm.grid(column=1, row=1)
+                self.newUserSubtitle.grid(column=0, row=2)
+                self.newUserForm.grid(column=1, row=2)
+                self.usernameChangeConfirm.grid(column=0, row=3, columnspan=2)
+            else:
+                self.passIncorrect = tk.Label(self, text="Password is incorrect", bg="red")
+                self.passIncorrect.grid(column=0, row=4, sticky="news", columnspan=2)
+        else:
+            self.passNotMatch = tk.Label(self, text="Passwords do not match", bg="red")
+            self.passNotMatch.grid(column=0, row=5, sticky="news", columnspan=2)
+
+    def confirmusernamechange(self):
+        if currentUser == self.currentUserForm.get():
+            if self.newUserForm.get() != " " and len(self.newUserForm.get()) > 0:
+                DataLocation = ("./usr/%s/AccountDetails.dat" % currentUser)
+                with open(DataLocation, "rb+") as UsrData:
+                    Details = pickle.load(UsrData)
+                    UsrData.close()
+                passvarHashed = Details[1]
+                usrVariables = [currentUser, passvarHashed, Wins, Loses, Draws, accountType, profilePicture]
+                with open(DataLocation, "wb+") as data:
+                    pickle.dump(usrVariables, data, 0)
+                    data.close()
+                os.rename("./usr/%s" % currentUser, "./usr/%s" % self.newUserForm.get())
+                global OldUsername
+                OldUsername = currentUser
+                self.inituserchange()
+            else:
+                newUserInvalid = tk.Label(self, text="New username cannot be empty", bg="red")
+                newUserInvalid.grid(column=0, row=4, sticky="news", columnspan=2)
+        else:
+            currentUserIncorrect = tk.Label(self, text="Current username is incorrect", bg="red")
+            currentUserIncorrect.grid(column=0, row=5, sticky="news", columnspan=2)
+
+    def inituserchange(self):
+        global currentUser
+        currentUser = self.newUserForm.get()
+        for widget in self.winfo_children():
+            widget.destroy()
+        self.title = tk.Label(self, text="Username Successfully changed", bg="white", font="TkFixedFont 15 bold")
+        self.subtitle = tk.Label(self, text="From %s --> %s" % (OldUsername, currentUser), bg="white", font="TkFixedFont 12 bold")
+        self.quitButton = tk.Button(self, text="Exit", bg="white", command=profilewindow.destroy)
+        self.title.grid(column=0, row=0)
+        self.subtitle.grid(column=0, row=1)
+        self.quitButton.grid(column=0, row=2)
+    
+    def changeaccountpassword(self):
+        print("placeholder")
         
 startwindow = tk.Tk()
 startMenu(startwindow).pack(fill="both", expand=True)
